test(cli): constrain nonce boundary test to unix
